feat: add the ability to change language and disable typing of command keys
