+++ virtual-keyboard/index.html
@@ -15,5 +15,7 @@
   <script type="module" src="script.js"></script>
   <script type="module" src="classes.js"></script>
   <script type="module" src="keys-config-parser.js"></script>
+  <script type="module" src="build-layout.js"></script>
+  <script type="module" src="key-codes.js"></script>
 </body>
 </html>
