+++ virtual-keyboard/key-codes.js
@@ -1,0 +1,11 @@
+export const keyCodes = [
+    "Backquote", "Digit1", "Digit2", "Digit3", "Digit4",  "Digit5",
+    "Digit6",  "Digit7",  "Digit8",  "Digit9", "Digit0", "Minus",
+    "Equal", "Backspace", "Tab", "KeyQ", "KeyW", "KeyE", "KeyR", "KeyT", "KeyY", "KeyU",
+    "KeyI", "KeyO", "KeyP", "BracketLeft", "BracketRight", "Delete", "CapsLock",
+    "KeyA", "KeyS", "KeyD", "KeyF", "KeyG", "KeyH", "KeyJ", "KeyK", "KeyL",
+    "Semicolon", "Quote", "Enter", "ShiftLeft", "KeyZ", "KeyX", "KeyC", "KeyV",
+    "KeyB", "KeyN", "KeyM", "Comma", "Period", "Slash", "ArrowUp", "ShiftRight",
+    "ControlLeft", "MetaLeft", "AltLeft", "Space", "AltRight", "ArrowLeft",
+    "ArrowDown", "ArrowRight", "ControlRight",
+]
